refactor(output): extract activity.ts + writer.ts — no circular deps, index.ts 1156→589 lines
